requirements.txt
discord.py>=2.4.0,<2.5.0
requests>=2.32.3,<2.33.0
python-dotenv>=1.0.1,<1.1.0
youtube-dl
yt-dlp
PyNaCl>=1.5.0,<1.6
